Fixed the broken Settings
JavaScript-based Font Size support is now implemented in the HTML file,
the settings are now stored as ints instead of doubles (since that
actually works), and a bug that prevented settings from ever being
saved is fixed

+++ Controller/GameMenuViewController.h
@@ -18,7 +18,7 @@
 	UISlider *fontSizeSlider;
 	UISlider *soundVolumeSlider;
 	UISlider *musicVolumeSlider;
-	double *fontSize, *soundVolume, *musicVolume;
+	int *fontSize, *soundVolume, *musicVolume;
 	
 	UIPopoverController *padPopoverController;
 }
@@ -31,7 +31,7 @@
 @property (nonatomic, retain) IBOutlet UISlider *soundVolumeSlider;
 @property (nonatomic, retain) IBOutlet UISlider *musicVolumeSlider;
 
-@property (nonatomic, assign) double *fontSize, *soundVolume, *musicVolume;
+@property (nonatomic, assign) int *fontSize, *soundVolume, *musicVolume;
 
 @property (nonatomic, retain) UIPopoverController *padPopoverController;
